RAG_System/requirements_txt.txt
# Core Dependencies
pandas>=2.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pyarrow>=12.0.0
python-dotenv>=1.0.0
numpy>=1.24.0

# Machine Learning & NLP
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0
scikit-learn>=1.3.0

# LangChain Ecosystem
langchain>=0.0.300
langchain-community>=0.0.10

# Vector Database
pgvector>=0.2.0

# Web Interface
gradio>=4.0.0

# Monitoring & Metrics
prometheus-client>=0.17.0
psutil>=5.9.0

# Optional Performance Enhancers
# numba>=0.57.0  # For numerical acceleration
# faiss-cpu>=1.7.4  # Alternative vector search

# Development Tools (optional)
# pytest>=7.0.0
# black>=23.0.0
# flake8>=6.0.0
# mypy>=1.0.0
# pre-commit>=3.0.0